+++ src/main/java/com/hangout/hangout/domain/auth/service/AuthService.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hangout.hangout.domain.auth.dto.request.LoginReqeust;
 import com.hangout.hangout.domain.auth.dto.request.SignUpRequest;
 import com.hangout.hangout.domain.auth.dto.response.AuthResponse;
-import com.hangout.hangout.domain.auth.entity.Token;
-import com.hangout.hangout.domain.auth.entity.TokenType;
 import com.hangout.hangout.domain.auth.repository.TokenRepository;
 import com.hangout.hangout.domain.user.entity.User;
 import com.hangout.hangout.domain.user.repository.UserRepository;
@@ -15,7 +13,6 @@ import com.hangout.hangout.global.exception.NotFoundException;
 import com.hangout.hangout.global.security.JwtService;
 import com.hangout.hangout.global.security.UserPrincipal;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -59,8 +56,8 @@ public class AuthService {
             .orElseThrow(() -> new NotFoundException(ResponseType.USER_NOT_EXIST_EMAIL));
         String jwtToken = jwtService.generateToken(UserPrincipal.create(user));
         String refreshToken = jwtService.generateRefreshToken(UserPrincipal.create(user));
-        revokeAllUserTokens(user);
-        saveUserToken(user, refreshToken);
+        jwtService.revokeAllUserTokens(user);
+        jwtService.saveUserToken(user, refreshToken);
         return createAuthResponse(jwtToken, refreshToken);
     }
 
@@ -69,29 +66,6 @@ public class AuthService {
             .accessToken(accessToken)
             .refreshToken(refreshToken)
             .build();
-    }
-
-    private void saveUserToken(User user, String jwtToken) {
-        Token token = Token.builder()
-            .user(user)
-            .token(jwtToken)
-            .tokenType(TokenType.BEARER)
-            .expired(false)
-            .revoked(false)
-            .build();
-        tokenRepository.save(token);
-    }
-
-    private void revokeAllUserTokens(User user) {
-        List<Token> validUserTokens = tokenRepository.findAllValidTokensByUserId(user.getId());
-        if (validUserTokens.isEmpty()) {
-            return;
-        }
-        validUserTokens.forEach(token -> {
-            token.setExpired(true);
-            token.setRevoked(true);
-        });
-        tokenRepository.saveAll(validUserTokens);
     }
 
     @Scheduled(fixedRate = 604800000) // 1 week
@@ -119,8 +93,8 @@ public class AuthService {
                 .orElseThrow();
             if (jwtService.isTokenValid(refreshToken, UserPrincipal.create(user))) {
                 String accessToken = jwtService.generateToken(UserPrincipal.create(user));
-                revokeAllUserTokens(user);
-                saveUserToken(user, accessToken);
+                jwtService.revokeAllUserTokens(user);
+                jwtService.saveUserToken(user, accessToken);
                 var authResponse = AuthResponse.builder()
                     .accessToken(accessToken)
                     .refreshToken(refreshToken)
